+++ releaseNotes.md
@@ -1,3 +1,9 @@
+# Version 0.3
+
+- Added inventory for following Entra ID components
+  - Enterprise Applications
+  - Managed Identities
+
 # Version 0.2
 
 - Increased max. number of results per query from 100 to 1000